+++ src/scripts/create_multi_wiki_qa.py
@@ -44,6 +44,7 @@ LANGUAGES = [
     "lv",
     "lt",
     "sk",
+    "uk",
 ]