+++ src/scripts/create_winogrande.py
@@ -45,6 +45,7 @@ LANGUAGES = [
     "pt",
     "sk",
     "sv",
+    "uk",
 ]